+++ packages/core/src/stream-loading/anchor-timestamp-extractor.ts
