+++ packages/core/src/stream-loading/log-syncer.ts
