+++ packages/core/src/stream-loading/state-manipulator.ts
